+++ CHANGELOG.md
@@ -4,3 +4,11 @@ Changelog
 ### 1.0.0
 -  First release was published.
 
+
+
+TODO
+=========
+- Tests for fragment utils
+- Tests for existing parsers
+- Add more parsers
+- Diagnostic/crash log